pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "celltype-cli"
version = "0.2.0"
description = "CellType CLI — An autonomous agent for drug discovery research"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "CellType Inc." },
]

dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "anthropic>=0.74.0",
    "openai>=1.0",
    "claude-agent-sdk>=0.1",
    "httpx>=0.27",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "python-dotenv>=1.0",
    "markdown>=3.5",
    "nbformat>=5.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
notebook = [
    "nbconvert>=7.0",
]
chemistry = [
    "rdkit>=2023.03",
]
biology = [
    "biopython>=1.81",
]
singlecell = [
    "scanpy>=1.9",
    "anndata>=0.10",
    "celltypist>=1.6",
    "cellxgene-census>=1.0",
    "tiledbsoma>=1.0",
]
ml = [
    "torch>=2.0",
    "transformers>=4.40",
    "fair-esm>=2.0",
]
analysis = [
    "seaborn>=0.13",
    "scikit-learn>=1.3",
]
api = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "duckdb>=1.0",
]
alpha = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "duckdb>=1.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
]
cloud = [
    "modal>=1.0",
    "pyjwt>=2.0",
]
all = [
    "rdkit>=2023.03",
    "biopython>=1.81",
    "scanpy>=1.9",
    "anndata>=0.10",
    "celltypist>=1.6",
    "cellxgene-census>=1.0",
    "tiledbsoma>=1.0",
    "torch>=2.0",
    "transformers>=4.40",
    "fair-esm>=2.0",
    "seaborn>=0.13",
    "scikit-learn>=1.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
ct = "ct.cli:entry"

[tool.hatch.build.targets.wheel]
packages = ["src/ct"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests hitting real APIs (use --run-e2e)",
    "docker: tests requiring Docker (use --run-e2e)",
    "e2e_matrix: optional live multi-prompt end-to-end matrix (set CT_RUN_E2E_MATRIX=1)",
    "api_smoke: optional live API smoke tests (set CT_RUN_API_SMOKE=1)",
]
